Logging of all request and errors
