+++ packages/graphql-api/package.json
@@ -49,7 +49,6 @@
     "topojson": "^3.0.2",
     "wikibase-sdk": "^7.0.8",
     "winston": "^3.11.0",
-    "winston-elasticsearch": "^0.17.4",
     "xml2js": "^0.4.23",
     "yaml": "^1.10.0"
   },
