--- packages/graphql-api/src/plugins/index.ts
@@ -1,2 +1,0 @@
-export { errorLoggingPlugin } from './errorLoggingPlugin';
-export { slowQueryLoggingPlugin } from './slowQueryLoggingPlugin';
